+++ common/src/main/java/de/maxhenkel/voicechat/Voicechat.java
@@ -17,7 +17,7 @@ public abstract class Voicechat {
 
     public static int COMPATIBILITY_VERSION = -1;
 
-    public static final Pattern GROUP_REGEX = Pattern.compile("^\\S[^\"\\n\\r\\t]{0,15}$");
+    public static final Pattern GROUP_REGEX = Pattern.compile("^[^\"\\n\\r\\t\\s][^\"\\n\\r\\t]{0,15}$");
 
     public void initialize() {
         CommonCompatibilityManager.INSTANCE = createCompatibilityManager();