+++ common/src/main/java/de/maxhenkel/voicechat/net/LeaveGroupPacket.java
@@ -1,0 +1,35 @@
+package de.maxhenkel.voicechat.net;
+
+import de.maxhenkel.voicechat.Voicechat;
+import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.resources.ResourceLocation;
+
+public class LeaveGroupPacket implements Packet<LeaveGroupPacket> {
+
+    public static final ResourceLocation LEAVE_GROUP = new ResourceLocation(Voicechat.MODID, "leave_group");
+
+    public LeaveGroupPacket() {
+
+    }
+
+    @Override
+    public ResourceLocation getIdentifier() {
+        return LEAVE_GROUP;
+    }
+
+    @Override
+    public int getID() {
+        return 8;
+    }
+
+    @Override
+    public LeaveGroupPacket fromBytes(FriendlyByteBuf buf) {
+        return this;
+    }
+
+    @Override
+    public void toBytes(FriendlyByteBuf buf) {
+
+    }
+
+}