+++ common/src/main/java/de/maxhenkel/voicechat/net/NetManager.java
@@ -11,15 +11,21 @@ public abstract class NetManager {
     public Channel<PlayerStatePacket> playerStateChannel;
     public Channel<PlayerStatesPacket> playerStatesChannel;
     public Channel<SecretPacket> secretChannel;
-    public Channel<SetGroupPacket> setGroupChannel;
     public Channel<RequestSecretPacket> requestSecretChannel;
+    public Channel<JoinGroupPacket> joinGroupChannel;
+    public Channel<CreateGroupPacket> createGroupChannel;
+    public Channel<LeaveGroupPacket> leaveGroupChannel;
+    public Channel<JoinedGroupPacket> joinedGroupChannel;
 
     public void init() {
         playerStateChannel = registerReceiver(PlayerStatePacket.class, true, true);
         playerStatesChannel = registerReceiver(PlayerStatesPacket.class, true, false);
         secretChannel = registerReceiver(SecretPacket.class, true, false);
-        setGroupChannel = registerReceiver(SetGroupPacket.class, true, false);
         requestSecretChannel = registerReceiver(RequestSecretPacket.class, false, true);
+        joinGroupChannel = registerReceiver(JoinGroupPacket.class, false, true);
+        createGroupChannel = registerReceiver(CreateGroupPacket.class, false, true);
+        leaveGroupChannel = registerReceiver(LeaveGroupPacket.class, false, true);
+        joinedGroupChannel = registerReceiver(JoinedGroupPacket.class, true, false);
     }
 
     public abstract <T extends Packet<T>> Channel<T> registerReceiver(Class<T> packetType, boolean toClient, boolean toServer);