+++ common/src/main/java/de/maxhenkel/voicechat/voice/server/PlayerStateManager.java
@@ -14,7 +14,7 @@ import java.util.concurrent.ConcurrentHashMap;
 
 public class PlayerStateManager {
 
-    private ConcurrentHashMap<UUID, PlayerState> states;
+    private final ConcurrentHashMap<UUID, PlayerState> states;
 
     public PlayerStateManager() {
         states = new ConcurrentHashMap<>();
@@ -22,8 +22,15 @@ public class PlayerStateManager {
         CommonCompatibilityManager.INSTANCE.onPlayerLoggedIn(this::notifyPlayer);
 
         CommonCompatibilityManager.INSTANCE.getNetManager().playerStateChannel.registerServerListener((server, player, handler, packet) -> {
+            PlayerState oldState = states.get(player.getUUID());
+
             PlayerState state = packet.getPlayerState();
             state.setGameProfile(player.getGameProfile());
+            if (oldState != null) {
+                state.setGroup(oldState.getGroup());
+            } else {
+                state.setGroup(null);
+            }
             states.put(player.getUUID(), state);
             broadcastState(server, state);
         });
@@ -48,6 +55,15 @@ public class PlayerStateManager {
     @Nullable
     public PlayerState getState(UUID playerUUID) {
         return states.get(playerUUID);
+    }
+
+    public static PlayerState defaultDisconnectedState(ServerPlayer player) {
+        return new PlayerState(false, true, player.getGameProfile());
+    }
+
+    public void setState(MinecraftServer server, UUID playerUUID, PlayerState state) {
+        states.put(playerUUID, state);
+        broadcastState(server, state);
     }
 
     public Collection<PlayerState> getStates() {